requirements.txt
asgiref
beautifulsoup4
certifi
crispy-bootstrap5
distlib
dj-database-url
Django
django-bootstrap4
django-crispy-forms
django-email
django-environ
django-filter
djangorestframework
EasyProcess
entrypoint2
filelock
Markdown
MouseInfo
mss
Pillow
pip
pipenv
platformdirs
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
pyscreenshot
PyScreeze
python-decouple
pytweening
pytz
setuptools
soupsieve
sqlparse
typing_extensions
tzdata
virtualenv